+++ src/lib.rs
@@ -373,6 +373,11 @@ impl JsValue {
     pub fn is_undefined(&self) -> bool {
         __wbindgen_is_undefined(self)
     }
+    /// Tests whether this JS value is `null` or `undefined`
+    #[inline]
+    pub fn is_null_or_undefined(&self) -> bool {
+        unsafe { __wbindgen_is_null_or_undefined(self.idx) }
+    }
 
     /// Tests whether the type of this JS value is `symbol`
     #[inline]
@@ -1143,7 +1148,7 @@ impl<T: TryFromJsValue> TryFromJsValue for Option<T> {
     type Error = T::Error;
 
     fn try_from_js_value(value: JsValue) -> Result<Self, Self::Error> {
-        if value.is_undefined() || value.is_null() {
+        if value.is_null_or_undefined() {
             Ok(None)
         } else {
             T::try_from_js_value(value).map(Some)
@@ -1295,6 +1300,7 @@ externs! {
     extern "C" {
         fn __wbindgen_object_clone_ref(idx: u32) -> u32;
         fn __wbindgen_object_drop_ref(idx: u32) -> ();
+        fn __wbindgen_is_null_or_undefined(idx: u32) -> bool;
 
         fn __wbindgen_describe(v: u32) -> ();
         fn __wbindgen_describe_cast(func: *const (), prims: *const ()) -> *const ();